Second try to fix BARSFE perf tests following Stride auth integration
