--- src/test/resources/data/allSortCodesNew.csv
@@ -1,2 +1,0 @@
-sortCode
-010004
