+++ src/test/resources/gatling.conf
@@ -66,7 +66,7 @@ gatling {
     #maxThreadLocalCharBufferSize = 200000  # Netty's default is 16k
   }
   ssl {
-    useOpenSsl = false                    # if OpenSSL should be used instead of JSSE (only the latter can be debugged with -Djava.net.debug=ssl)
+    #useOpenSsl = true                    # if OpenSSL should be used instead of JSSE (only the latter can be debugged with -Djava.net.debug=ssl)
     #useOpenSslFinalizers = false         # if OpenSSL contexts should be freed with Finalizer or if using RefCounted is fine
     #handshakeTimeout = 10000             # TLS handshake timeout in millis
     #useInsecureTrustManager = true       # Use an insecure TrustManager that trusts all server certificates
