+++ src/test/resources/journeys.conf
@@ -23,10 +23,10 @@ journeys {
 
     # The load is in journeys per second. Put here the load you are going to have at the peak.
     # There is no need to put a higher value at this point. Use prerftest.loadPercentage in application.conf instead
-    load = 0
+    load = 150
 
     # This points to a csv file with the data you need to inject in the journey. [More here](https://github.com/hmrc/performance-test-runner#step-4-configure-the-user-feeder)
-    feeder = data/allSortCodesNew.csv
+    feeder = data/allSortCodes.csv
 
     # The parts your journey is made of. A part is made one or more requests.
     parts = [
@@ -131,13 +131,11 @@ journeys {
 
 }
 
-
 # Default behaviour is to run all journeys. If that is not what you need you can specify the list of journeys to run
 journeysToRun = [
-  bars-frontend-simulation
-//  bars-business-simulation-stubbed,
-//  bars-individual-simulation-stubbed,
-//  bars-validate-simulation
+  bars-business-simulation-stubbed,
+  bars-individual-simulation-stubbed,
+  bars-validate-simulation
 ]
 
 # You can specify the same list of journeys via environment variables:
