+++ src/test/resources/services-local.conf
@@ -33,12 +33,12 @@
 
 
 services {
-
   bank-account-reputation-frontend.protocol = http
   bank-account-reputation-frontend.host = "localhost"
   bank-account-reputation-frontend.port = 9786
   bank-account-reputation.protocol = http
   bank-account-reputation.host = "localhost"
   bank-account-reputation.port = 9871
-
+  stride-stub.port = 9043
+  stride-auth.port = 9041
 }
