+++ src/test/resources/services.conf
@@ -32,11 +32,24 @@
 # helloworld-service.host = internal.helloworld-service.co.uk
 
 services {
-    bank-account-reputation-frontend.protocol = https
-    bank-account-reputation-frontend.host = "www.staging.tax.service.gov.uk"
-    bank-account-reputation-frontend.port = 443
-
-    # This has been set to override the baseURL which is being ignore in the jkenkins config.
-    bank-account-reputation.protocol = https
-    bank-account-reputation.host = "bank-account-reputation.protected.mdtp"
+    bank-account-reputation-frontend {
+        protocol = https
+        host = "admin.staging.tax.service.gov.uk"
+        port = 443
+    }
+    # This has been set to override the baseURL which is being ignored in the Jenkins config.
+    bank-account-reputation {
+        protocol = https
+        host = "bank-account-reputation.protected.mdtp"
+    }
+    stride-stub {
+        protocol = https
+        host =  admin.staging.tax.service.gov.uk
+        port = 443
+    }
+    stride-auth {
+        protocol = https
+        host =  admin.staging.tax.service.gov.uk
+        port = 443
+    }
 }
